+++ src/main/java/org/folio/entitlement/integration/kafka/EntitlementEventPublisher.java
@@ -19,6 +19,7 @@ public class EntitlementEventPublisher {
    */
   public void publish(EntitlementEvent event) {
     var entitlementTopic = getEnvTopicName("entitlement");
-    kafkaEventPublisher.send(entitlementTopic, event.getTenantName() + "_" + event.getModuleId(), event);
+    kafkaEventPublisher.send(entitlementTopic, event.getTenantName() + "_" + event.getModuleId(),
+      event.getTenantName(), event);
   }
 }